+++ markov_move.py
@@ -21,8 +21,8 @@ def conjugation_markov_move(braid: torch.Tensor, j: int, k: int) -> torch.Tensor
     j = 1, 2, ..., max(abs(braid))
     """
     # Calculate the terms to add
-    term1 = torch.tensor([(-1) ** k * j], dtype=torch.float16)
-    term2 = torch.tensor([(-1) ** (k + 1) * j], dtype=torch.float16)
+    term1 = torch.tensor([(-1) ** k * j], dtype=torch.float)
+    term2 = torch.tensor([(-1) ** (k + 1) * j], dtype=torch.float)
     # Update braid: braid ← [(−1)^k j] + braid + [(−1)^(k+1) j]
     return torch.cat((term1, braid, term2), dim=0)
 
@@ -44,7 +44,7 @@ def new_strand_markov_move(braid: torch.Tensor, k: int) -> torch.Tensor:
     # Calculate the term to add
     term = (-1) ** k * (max_abs_braid(braid) + 1)
     # Update braid: braid ← braid + [(−1)^k (max(abs(braid))+1)]
-    new_element = torch.tensor([term], dtype=torch.float16)
+    new_element = torch.tensor([term], dtype=torch.float)
     return torch.cat((braid, new_element))
