+++ markov_move_test.py
@@ -4,7 +4,7 @@ from markov_move import random_conjugation_markov_move, random_new_strand_markov
 
 
 def test_conjugation_markov_move():
-    braid = torch.tensor([-1, 2, 4, -3, 1], dtype=torch.float16)
+    braid = torch.tensor([-1, 2, 4, -3, 1], dtype=torch.float)
     new_braid = random_conjugation_markov_move(braid.clone())
     assert len(new_braid) == len(braid) + 2
     assert new_braid[0] == - 1 * new_braid[-1]
@@ -14,7 +14,7 @@ def test_conjugation_markov_move():
 
 
 def test_new_strand_markov_move():
-    braid = torch.tensor([-1, 2, 4, -3, 1], dtype=torch.float16)
+    braid = torch.tensor([-1, 2, 4, -3, 1], dtype=torch.float)
     new_braid = random_new_strand_markov_move(braid.clone())
     assert len(new_braid) == len(braid) + 1
     assert new_braid[:-1].equal(braid)
